+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 - Support for [scoped entities](https://cap.cloud.sap/docs/cds/cdl#scoped-names)
 
+### Fixed
+- Inline enums are now available during runtime as well
+
 ## Version 0.14.0 - 2023-12-13
 ### Added
 - Entities that are database views now also receive typings